redesigned mcu delay for AVR
-delay for AVR based on avr/delay.h was blocking the process. Replaced by internal version based on RTC clock

+++ uCNC/mcu.h
@@ -112,7 +112,7 @@ uint32_t mcu_millis();
 
 #ifndef mcu_delay_ms
 //Custom delay function
-void mcu_delay_ms(uint16_t miliseconds);
+void mcu_delay_ms(uint32_t miliseconds);
 #endif
 
 //Non volatile memory
